attempt to rename folder
