+++ src/test/java/dev/coms4156/project/druginteraction/FirebaseServiceUnitTests.java
@@ -1,4 +1,4 @@
-package dev.coms4156.project.individualproject;
+package dev.coms4156.project.druginteraction;
 
 import java.util.Map;
 import java.util.HashMap;
@@ -19,6 +19,8 @@ import org.junit.jupiter.api.TestMethodOrder;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+
+// import dev.coms4156.project.druginteraction.FirebaseService;
 
 /**
  * Unit tests for the FirebaseService and FirebaseConfig classes.
